+++ sdk/textanalytics/ai-text-analytics/test/public/analyzeBatch.spec.ts
@@ -820,7 +820,7 @@ matrix([["APIKey", "AAD"]] as const, async (authMethod: AuthMethod) => {
           poller.onProgress((state) => {
             assert.ok(state.createdOn, "createdOn is undefined!");
             assert.ok(state.expiresOn, "expiresOn is undefined!");
-            assert.ok(state.lastModifiedOn, "lastModifiedOn is undefined!");
+            assert.ok(state.modifiedOn, "modifiedOn is undefined!");
             assert.ok(state.status, "status is undefined!");
             assert.ok(state.operationId, "operationId is undefined!");
             assert.equal(state.displayName, "testJob");